% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{build_geojson_feature}
\alias{build_geojson_feature}
\title{Build a single GeoJSON feature}
\usage{
build_geojson_feature(cleaned_coords, item)
}
\arguments{
\item{cleaned_coords}{A \code{character} vector of length 1 with the fixed coordinates.}

\item{item}{A \code{list} with the item data.}
}
\value{
A GeoJSON feature in a \code{character} vector of length 1.
}
\description{
This function creates a feature with the fixed coordinates and includes
the "toujours_signaler" field from the item as a property.
}
\keyword{internal}
