% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fix_brackets_iterative}
\alias{fix_brackets_iterative}
\title{Function to adjust the bracket levels of the string}
\usage{
fix_brackets_iterative(s)
}
\arguments{
\item{s}{a \code{character} vector of length 1 with the coordinates string to be fixed.}
}
\value{
a \code{character} vector of length 1 with the fixed coordinates string.
}
\description{
Function to adjust the bracket levels of the string
}
\keyword{internal}
