% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{js_coords_to_geojson}
\alias{js_coords_to_geojson}
\title{Convert JavaScript encoded coordinates to \code{GeoJSON} comliant string}
\usage{
js_coords_to_geojson(item)
}
\arguments{
\item{item}{A \code{list} with the item data.}
}
\value{
An \code{sf} spatial object.
}
\description{
Convert JavaScript encoded coordinates to \code{GeoJSON} comliant string
}
\keyword{internal}
