% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiger_mosquito_colonisation_in_france.R
\name{get_tiger_mosquito_colonisation_in_france}
\alias{get_tiger_mosquito_colonisation_in_france}
\title{Get latest tiger mosquito colonisation data in France from the online map}
\usage{
get_tiger_mosquito_colonisation_in_france()
}
\value{
An \code{sf} object with the tiger mosquito colonisation data in France. All polygons that exist on the map are covering the territories where the tiger mosquito (Aedes Albopictus) has alreadt been detected in France.
}
\description{
Get latest tiger mosquito colonisation data in France from the online map at \url{https://signalement-moustique.anses.fr/signalement_albopictus/colonisees}.
}
\examples{
\donttest{
# get the latest data
x <- get_tiger_mosquito_colonisation_in_france()

# save the data to GeoPackage file with `sf`
library(sf)
if(interactive()) {
st_write(x, "tiger_mosquito_colonisation_in_france.gpkg")
}
}

}
